# store-gateway: reduce overhead in fetching postings from cache