--- a/pkg/storegateway/indexcache/inmemory.go
+++ b/pkg/storegateway/indexcache/inmemory.go
@@ -310,7 +310,7 @@ func (labelsMapResult) Close() error { return nil }
 
 // FetchMultiPostings fetches multiple postings - each identified by a label -
 // and returns a map containing cache hits, along with a list of missing keys.
-func (c *InMemoryIndexCache) FetchMultiPostings(ctx context.Context, userID string, blockID ulid.ULID, keys []labels.Label) (_ Result[labels.Label], misses []labels.Label) {
+func (c *InMemoryIndexCache) FetchMultiPostings(ctx context.Context, userID string, blockID ulid.ULID, keys []labels.Label) (_ Result[labels.Label]) {
 	hits := labelsMapResult{}
 
 	for _, key := range keys {
@@ -318,11 +318,9 @@ func (c *InMemoryIndexCache) FetchMultiPostings(ctx context.Context, userID stri
 			hits[key] = b
 			continue
 		}
-
-		misses = append(misses, key)
 	}
 
-	return hits, misses
+	return hits
 }
 
 // StoreSeriesForRef sets the series identified by the ulid and id to the value v,