… CHANGELOG.md
… pkg/storegateway/bucket_test.go
… pkg/storegateway/indexcache/inmemory_test.go
… pkg/storegateway/indexcache/remote_test.go
… pkg/storegateway/indexcache/tracing.go
… pkg/storegateway/series_refs_test.go